Blender addon for additively mirroring shape keys
Small addon that adds two extra buttons to the Shape Key Specials Menu (shape key context menu) that are like "Mirror Shape Key" and "Mirror Shape Key (Topology)", but add the mirrored movement to the current shape. This makes it easy to take a shape key for one side of a model and then additively mirror it onto itself so that the shape key now affects both sides.